+++ example/example.js
@@ -17,6 +17,7 @@ const metadata = {
   language: 'en',
   description: 'A test book.',
   contents: 'Chapters',
+  embedCover: true,
   showContents: true,
   source: 'http://www.kcartlidge.com',
   images: ['test/hat.png']